+++ .gitignore
@@ -2,3 +2,4 @@
 
 .project
 .settings
+.idea
