+++ src/QueueConsumer.php
@@ -6,7 +6,8 @@ interface QueueConsumer
     /**
      * Consumes a message
      * 
-     * @param mixed $message
+     * @param  mixed $message
+     * @return mixed|null|void
      */
     public function consume($message);
 }
